fix: resolve schedule page layout and authorization issues
Comprehensive fix for schedule management dashboard addressing:

## Layout Issue Fix:
- Converted schedule page from client to server component pattern
- Created SchedulesClient component for interactive functionality
- Now matches layout pattern of other protected pages (chat, assistant-architect)
- Removed container styling that prevented proper sidebar display

## Authorization Issue Fix:
- Reverted tool access checks from 'architect' to 'assistant-architect'
- Updated API routes: /api/schedules, /api/schedules/[id]
- Updated server actions: schedule-actions.ts
- Fixed navigation migration: 036-add-schedule-navigation.sql

## Technical Changes:
- /app/(protected)/schedules/page.tsx - Now server component with auth check
- /app/(protected)/schedules/_components/schedules-client.tsx - Client logic moved here
- All tool access validation now uses correct 'assistant-architect' identifier
- Maintains same functionality with proper layout and permissions

Resolves 'Forbidden' error and missing sidebar layout issues.
