+++ app/api/schedules/[id]/route.ts
@@ -34,10 +34,10 @@ export async function GET(request: NextRequest, { params }: RouteParams) {
       )
     }
 
-    // Check if user has access to the architect tool
-    const hasAccess = await hasToolAccess(session.sub, "architect")
+    // Check if user has access to the assistant-architect tool
+    const hasAccess = await hasToolAccess(session.sub, "assistant-architect")
     if (!hasAccess) {
-      log.warn("Forbidden - User lacks architect access")
+      log.warn("Forbidden - User lacks assistant-architect access")
       timer({ status: "error", reason: "forbidden" })
       return NextResponse.json(
         { isSuccess: false, message: "Forbidden" },
@@ -123,10 +123,10 @@ export async function PUT(request: NextRequest, { params }: RouteParams) {
       )
     }
 
-    // Check if user has access to the architect tool
-    const hasAccess = await hasToolAccess(session.sub, "architect")
+    // Check if user has access to the assistant-architect tool
+    const hasAccess = await hasToolAccess(session.sub, "assistant-architect")
     if (!hasAccess) {
-      log.warn("Forbidden - User lacks architect access")
+      log.warn("Forbidden - User lacks assistant-architect access")
       timer({ status: "error", reason: "forbidden" })
       return NextResponse.json(
         { isSuccess: false, message: "Forbidden" },
@@ -254,10 +254,10 @@ export async function DELETE(request: NextRequest, { params }: RouteParams) {
       )
     }
 
-    // Check if user has access to the architect tool
-    const hasAccess = await hasToolAccess(session.sub, "architect")
+    // Check if user has access to the assistant-architect tool
+    const hasAccess = await hasToolAccess(session.sub, "assistant-architect")
     if (!hasAccess) {
-      log.warn("Forbidden - User lacks architect access")
+      log.warn("Forbidden - User lacks assistant-architect access")
       timer({ status: "error", reason: "forbidden" })
       return NextResponse.json(
         { isSuccess: false, message: "Forbidden" },
